+++ tests/TestCase/GlobalFunctionsTest.php
@@ -17,7 +17,6 @@ namespace Tools\Test;
 
 use App\ExampleChildClass;
 use App\ExampleClass;
-use LogicException;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 use Tools\TestSuite\TestTrait;