+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -22,14 +22,9 @@ use App\ExampleChildClass;
 use App\ExampleClass;
 use App\SkipTestCase;
 use ArrayIterator;
-use BadMethodCallException;
-use ErrorException;
-use Exception;
 use GdImage;
 use IteratorAggregate;
 use PHPUnit\Framework\AssertionFailedError;
-use PHPUnit\Framework\Error\Deprecated;
-use PHPUnit\Framework\Error\Notice;
 use PHPUnit\Framework\Exception as PHPUnitException;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
@@ -156,142 +151,6 @@ class TestTraitTest extends TestCase
 
     /**
      * @test
-     * @uses \Tools\TestSuite\TestTrait::assertException()
-     */
-    public function testAssertException(): void
-    {
-        $this->TestCase->assertException(function () {
-            throw new Exception();
-        });
-        $this->TestCase->assertException(function () {
-            throw new Exception('right exception message');
-        }, Exception::class, 'right exception message');
-
-        //It correctly ignores deprecations
-        $this->TestCase->assertException(function () {
-            if (function_exists('Cake\Core\deprecationWarning')) {
-                \Cake\Core\deprecationWarning('1.0', 'This is a deprecation!');
-            } else {
-                \deprecationWarning('This is a deprecation');
-            }
-            throw new ErrorException('This is an error exception');
-        }, ErrorException::class, 'This is an error exception');
-
-        //No exception throw
-        try {
-            $this->TestCase->assertException('time');
-        } catch (AssertionFailedError $e) {
-            $this->assertSame('Expected exception `Exception`, but no exception throw', $e->getMessage());
-        } finally {
-            if (!isset($e)) {
-                self::fail();
-            }
-            unset($e);
-        }
-
-        //The comparison is strict, it does not consider parent classes (`ErrorException` != `Exception`)
-        try {
-            $this->TestCase->assertException(function () {
-                throw new ErrorException();
-            });
-        } catch (AssertionFailedError $e) {
-            $this->assertStringStartsWith('Expected exception `Exception`, unexpected type `ErrorException`', $e->getMessage());
-        } finally {
-            if (!isset($e)) {
-                self::fail();
-            }
-            unset($e);
-        }
-
-        //No existing exception or invalid exception class
-        foreach (['noExistingException', stdClass::class] as $class) {
-            try {
-                $this->TestCase->assertException(function () {
-                    throw new Exception();
-                }, $class);
-            } catch (AssertionFailedError $e) {
-                $this->assertSame('Class `' . $class . '` is not a throwable or does not exist', $e->getMessage());
-            } finally {
-                if (!isset($e)) {
-                    self::fail();
-                }
-                unset($e);
-            }
-        }
-
-        //Unexpected exception type
-        try {
-            $this->TestCase->assertException(function () {
-                throw new Exception();
-            }, BadMethodCallException::class);
-        } catch (AssertionFailedError $e) {
-            $this->assertStringStartsWith('Expected exception `' . BadMethodCallException::class . '`, unexpected type `Exception`', $e->getMessage());
-        } finally {
-            if (!isset($e)) {
-                self::fail();
-            }
-            unset($e);
-        }
-
-        //Wrong exception message
-        try {
-            $this->TestCase->assertException(function () {
-                throw new Exception('Wrong');
-            }, Exception::class, 'Right');
-        } catch (AssertionFailedError $e) {
-            $this->assertStringStartsWith('Expected message exception `Right`, unexpected message `Wrong`', $e->getMessage());
-        } finally {
-            if (!isset($e)) {
-                self::fail();
-            }
-            unset($e);
-        }
-
-        //Expected exception message, but no message
-        try {
-            $this->TestCase->assertException(function () {
-                throw new Exception();
-            }, Exception::class, 'Right');
-        } catch (AssertionFailedError $e) {
-            $this->assertStringStartsWith('Expected message exception `Right`, but no message for the exception', $e->getMessage());
-        } finally {
-            if (!isset($e)) {
-                self::fail();
-            }
-            unset($e);
-        }
-    }
-
-    /**
-     * Test for `assertException` with a deprecation (it can't assert deprecations)
-     * @test
-     * @uses \Tools\TestSuite\TestTrait::assertException()
-     */
-    public function testAssertExceptionWithDeprecation(): void
-    {
-        $this->skipIf(!class_exists(Deprecated::class));
-
-        //Can't assert deprecations
-        try {
-            $this->TestCase->assertException(function () {
-                if (function_exists('Cake\Core\deprecationWarning')) {
-                    \Cake\Core\deprecationWarning('1.0', 'This is a deprecation!');
-                } else {
-                    \deprecationWarning('This is a deprecation');
-                }
-            }, Deprecated::class);
-        } catch (Notice $e) {
-            $this->assertSame('You cannot use `assertException()` for deprecations, use instead `assertDeprecated()`', $e->getMessage());
-        } finally {
-            if (!isset($e)) {
-                self::fail();
-            }
-            unset($e);
-        }
-    }
-
-    /**
-     * @test
      * @uses \Tools\TestSuite\TestTrait::assertFileExtension()
      */
     public function testAssertFileExtension(): void
@@ -330,19 +189,40 @@ class TestTraitTest extends TestCase
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertIsArrayNotEmpty()
      */
-    public function testAssertIsArrayNotEmpty(): void
+    public function testAssertIsArrayNotEmptyWithEmptyArray(): void
     {
-        $this->TestCase->assertIsArrayNotEmpty(['value']);
+        $this->expectException(ExpectationFailedException::class);
+        $this->assertIsArrayNotEmpty([]);
+    }
 
-        foreach ([
-            [],
-            [[]],
-            [false],
-            [null],
-            [''],
-        ] as $array) {
-            $this->assertException(fn() => $this->TestCase->assertIsArrayNotEmpty($array), ExpectationFailedException::class);
-        }
+    /**
+     * @test
+     * @uses \Tools\TestSuite\TestTrait::assertIsArrayNotEmpty()
+     */
+    public function testAssertIsArrayNotEmptyWithBoolean(): void
+    {
+        $this->expectException(ExpectationFailedException::class);
+        $this->assertIsArrayNotEmpty(false);
+    }
+
+    /**
+     * @test
+     * @uses \Tools\TestSuite\TestTrait::assertIsArrayNotEmpty()
+     */
+    public function testAssertIsArrayNotEmptyWithNull(): void
+    {
+        $this->expectException(ExpectationFailedException::class);
+        $this->assertIsArrayNotEmpty(null);
+    }
+
+    /**
+     * @test
+     * @uses \Tools\TestSuite\TestTrait::assertIsArrayNotEmpty()
+     */
+    public function testAssertIsArrayNotEmptyWithString(): void
+    {
+        $this->expectException(ExpectationFailedException::class);
+        $this->assertIsArrayNotEmpty('');
     }
 
     /**